Integrated Sim's EventActivity

+++ app/src/main/java/com/example/customcalendar/AdapterPlan.kt
@@ -1,5 +1,6 @@
 package com.example.customcalendar
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.graphics.Color
 import android.util.Log
@@ -47,7 +48,33 @@ class AdapterPlan(val ids :MutableList<Int>, val titles: MutableList<String>, va
         tv.text = simpleDateFormat.format(selectedDate)
 
         val tv2 = holder.layout.findViewById<TextView>(R.id.planContent)
-        tv2.text = contents[position]
+
+        var tmpContent:String = contents[position]
+        tmpContent = tmpContent.replace("happy", "\uD83D\uDE04")
+        tmpContent = tmpContent.replace("smoking", "\uD83D\uDEAC")
+        tmpContent = tmpContent.replace("beer", "\uD83C\uDF7A")
+        tmpContent = tmpContent.replace("wine", "\uD83C\uDF77")
+        tmpContent = tmpContent.replace("cocktail", "\uD83C\uDF78")
+        tmpContent = tmpContent.replace("coffee", "☕")
+        tmpContent = tmpContent.replace("cake", "\uD83C\uDF70")
+        tmpContent = tmpContent.replace("money", "\uD83D\uDCB0")
+//
+//        when (mdContent) {
+//            "happy" -> mdContent = "\uD83D\uDE04" // 😄
+//            "smoking" -> mdContent = "\uD83D\uDEAC" // 🚬 (담배 이모티콘)
+//            "beer" -> mdContent = "\uD83C\uDF7A" // 🍺 (맥주 이모티콘)
+//            "wine" -> mdContent = "\uD83C\uDF77" // 🍷 (와인 이모티콘)
+//            "cocktail" -> mdContent = "\uD83C\uDF78" // 🍸 (칵테일 이모티콘)
+//            "coffee" -> mdContent = "\u2615" // ☕ (커피 이모티콘)
+//            "cake" -> mdContent = "\uD83C\uDF70" // 🍰 (케이크 이모티콘)
+//            "money" -> mdContent = "\uD83D\uDCB0" // 💰 (돈 이모티콘)
+//            else -> {
+//                // 기본적으로 입력된 텍스트 그대로 사용
+//            }
+//        }
+
+        tv2.text = tmpContent
+
 
         holder.layout.setOnClickListener {
             // Toast.makeText(holder.layout.context, tv.text.toString(), Toast.LENGTH_SHORT).show()
@@ -66,18 +93,32 @@ class AdapterPlan(val ids :MutableList<Int>, val titles: MutableList<String>, va
         }
 
         btnTrash.setOnClickListener {
-            val id = ids[position].toString()
-            val db = MyDb(it.context, null)
+            val builder = AlertDialog.Builder(it.context)
+            builder.setTitle("${tv.text.toString()} 삭제")
+            builder.setMessage("정말 삭제 하시겠습니까?")
 
-            if (db.deleteEvent(id)) {
-                Toast.makeText(it.context, "이벤트가 삭제되었습니다.", Toast.LENGTH_SHORT).show()
-            } else {
-                Toast.makeText(it.context, "이벤트 삭제 실패.", Toast.LENGTH_SHORT).show()
+            builder.setPositiveButton("삭제") { dialog, which ->
+                val id = ids[position].toString()
+                val db = MyDb(it.context, null)
+
+                if (db.deleteEvent(id)) {
+                    Toast.makeText(it.context, "이벤트가 삭제되었습니다.", Toast.LENGTH_SHORT).show()
+                } else {
+                    Toast.makeText(it.context, "이벤트 삭제 실패.", Toast.LENGTH_SHORT).show()
+                }
+
+                val mainActivity = Intent(it.context,MainActivity::class.java)
+                it.context.startActivity(mainActivity)
             }
 
-            val mainActivity = Intent(it.context,MainActivity::class.java)
-            it.context.startActivity(mainActivity)
+            builder.setNegativeButton("취소") { dialog, which ->
+                Toast.makeText(it.context, android.R.string.no, Toast.LENGTH_SHORT).show()
+            }
 
+//            builder.setNeutralButton("Maybe") { dialog, which ->
+//                Toast.makeText(it.context,"Maybe", Toast.LENGTH_SHORT).show()
+//            }
+            builder.show()
         }
 
         btnModify.setOnClickListener {
